entrenamiento/requirements.txt
openai-whisper 
jiwer 
pydub 
torch --index-url https://download.pytorch.org/whl/cpu 
